backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.4
google-genai==1.7.0
python-dotenv==1.0.1
pydantic==2.10.4
